Create iZEBot(2).ring
place this file in a seperate folder. The file will be created in that folder once it has been parsed. For now its outputting just a random message into the file but it can be modified to work with the tokens.  The parse tree is missing parts for 2 to 4 keys